Moved secured enum wrappers to their enum file
