+++ src/components/partnership/dto/partnership-agreement-status.enum.ts
@@ -1,4 +1,5 @@
-import { EnumType, makeEnum } from '~/common';
+import { ObjectType } from '@nestjs/graphql';
+import { EnumType, makeEnum, SecuredEnum } from '~/common';
 
 export type PartnershipAgreementStatus = EnumType<
   typeof PartnershipAgreementStatus
@@ -7,3 +8,10 @@ export const PartnershipAgreementStatus = makeEnum({
   name: 'PartnershipAgreementStatus',
   values: ['NotAttached', 'AwaitingSignature', 'Signed'],
 });
+
+@ObjectType({
+  description: SecuredEnum.descriptionFor('a partnership agreement status'),
+})
+export abstract class SecuredPartnershipAgreementStatus extends SecuredEnum(
+  PartnershipAgreementStatus,
+) {}
