+++ src/components/partnership/dto/partnership.dto.ts
@@ -8,7 +8,6 @@ import {
   Secured,
   SecuredBoolean,
   SecuredDateNullable,
-  SecuredEnum,
   SecuredProps,
   Sensitivity,
   SensitivityField,
@@ -20,23 +19,8 @@ import { ChangesetAware } from '../../changeset/dto';
 import { Organization } from '../../organization/dto';
 import { SecuredPartnerTypes } from '../../partner/dto';
 import { IProject } from '../../project/dto';
-import { FinancialReportingType } from './financial-reporting-type.enum';
-import { PartnershipAgreementStatus } from './partnership-agreement-status.enum';
-
-@ObjectType({
-  description: SecuredEnum.descriptionFor('a partnership agreement status'),
-})
-export abstract class SecuredPartnershipAgreementStatus extends SecuredEnum(
-  PartnershipAgreementStatus,
-) {}
-
-@ObjectType({
-  description: SecuredEnum.descriptionFor('partnership funding type'),
-})
-export abstract class SecuredFinancialReportingType extends SecuredEnum(
-  FinancialReportingType,
-  { nullable: true },
-) {}
+import { SecuredFinancialReportingType } from './financial-reporting-type.enum';
+import { SecuredPartnershipAgreementStatus } from './partnership-agreement-status.enum';
 
 @RegisterResource({ db: e.Partnership })
 @ObjectType({
