+++ src/components/user/dto/user-status.enum.ts
@@ -1,7 +1,13 @@
-import { EnumType, makeEnum } from '~/common';
+import { ObjectType } from '@nestjs/graphql';
+import { EnumType, makeEnum, SecuredEnum, SecuredProperty } from '~/common';
 
 export type UserStatus = EnumType<typeof UserStatus>;
 export const UserStatus = makeEnum({
   name: 'UserStatus',
   values: ['Active', 'Disabled'],
 });
+
+@ObjectType({
+  description: SecuredProperty.descriptionFor('a user status'),
+})
+export abstract class SecuredUserStatus extends SecuredEnum(UserStatus) {}
